fix: make integration test helpers warning-free
