+++ tests/v0_1_public_facade_tests.rs
@@ -6,6 +6,7 @@ use character_memory::{
 };
 use uuid::Uuid;
 
+#[path = "support/basic.rs"]
 mod test_utils;
 
 #[tokio::test]
